+++ src/main/java/util/URLMapper.java
@@ -1,12 +1,11 @@
 package util;
 
 import controller.UserController;
+import webserver.HttpRequest;
+import webserver.HttpResponse;
 
-import java.io.OutputStream;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.BiConsumer;
-import java.util.function.Consumer;
 import java.util.function.Function;
 
 public class URLMapper {
