+++ src/main/java/webserver/HttpRequest.java
@@ -1,4 +1,6 @@
-package util;
+package webserver;
+
+import util.URIParser;
 
 import java.io.BufferedReader;
 import java.io.IOException;
