+++ src/main/java/webserver/HttpResponse.java
@@ -1,4 +1,4 @@
-package util;
+package webserver;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
